% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_utils.R
\name{add_lat_long}
\alias{add_lat_long}
\title{Toevoegen van latidute en longitude}
\usage{
add_lat_long(data, x_coord = "x", y_coord = "y")
}
\arguments{
\item{data}{Dataframe waar latitude en longitude aan toegevoegd worden}

\item{x_coord}{Character. Kolomnaam van de x-coordinaat in RD-stelsel (EPSG:28992). Default is \code{"x"}}

\item{y_coord}{Character. Kolomnaam van de y-coordinaat in RD-stelsel (EPSG:28992). Default is \code{"y"}}
}
\value{
Het input dataframe met een kolom \code{long} en \code{lat} toegevoegd.
}
\description{
De functie voegt de latitude en longitude (WGS84) toe aan een dataframe op basis van RD-coordinaten
}
\examples{
\dontrun{

meetpunten \%>\% add_lat_long(x_coord = "x", y_coord = "y")
}

}
