% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functiesJT.R
\name{clean_vector}
\alias{clean_vector}
\title{Create a single clean vector}
\usage{
clean_vector(vector)
}
\arguments{
\item{vector}{A vector that needs 'cleaning'}
}
\value{
Returns a single clean vector
}
\description{
A helper function for the basic \code{c()} function to remove names and create a singular vector.
This can be useful if you select a single column from a dataframe.
}
