% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functiesJT.R
\name{knmi_neerslag_dag}
\alias{knmi_neerslag_dag}
\title{Neerslaggegevens}
\usage{
knmi_neerslag_dag(knmistation = "344")
}
\arguments{
\item{knmistation}{het nummer van het KNMI-station. Via de website zijn de nummers te vinden}
}
\value{
Een dataframe met twee kolommen datum en neerslag in mm
}
\description{
Deze functie haalt de neerslaggegevens op van een specifiek KNMI-station. De functie maakt gebruik van \code{\link{knmi_dag_ruw}}.
Voor meer toelichting kan die functie worden geraadpleegd.
}
