% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knmi_stations}
\alias{knmi_stations}
\title{Locaties KNMI-stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 5 columns.}
\usage{
knmi_stations
}
\description{
Bevat een dataframe met de locaties van alle KNMI-stations
}
\keyword{datasets}
