% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi_gegevens.R
\name{opschonen_knmi}
\alias{opschonen_knmi}
\title{Opschonen knmi data}
\usage{
opschonen_knmi(ruwe_data)
}
\arguments{
\item{ruwe_data}{Een dataframe met ruwe knmi dag data zoals gecreeerd door \code{\link[=knmi_dag_ruw]{knmi_dag_ruw()}}}
}
\value{
Een dataframe met begrijpelijke Nederlandse namen en logische eenheden
}
\description{
Deze functie schoont de KNMI-data die worden verkregen met \code{\link[=knmi_dag_ruw]{knmi_dag_ruw()}} op en geeft ze
begrijpelijke Nederlandse namen
}
\section{Eenheden}{

\itemize{
\item Windsnelheden zijn gemeten in m/s
\item Temperatuur is gemeten in graden Celsius
\item Straling is gemeten in J/cm2
\item Neerslag is gemeten is gemeten in mm
\item Luchtdruk is gemeten in hPa
\item Vochtigheid is gemeten in relatieve luchtvochtigheid
\item Verdamping is gemeten in mm
}
}

\examples{
\dontrun{

knmi_dag_ruw() \%>\% opschonen_knmi()
}
}
