% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_packcircles.R
\name{stat_packcircles_label}
\alias{stat_packcircles_label}
\title{Stat - Packcircles Label}
\usage{
stat_packcircles_label(mapping = NULL, data = NULL, geom = "text",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The mapping of aesthetics to use for this stat. Inherits from \code{ggplot()}}

\item{data}{The data to use for this stat. Inherits from \code{ggplot()}}

\item{geom}{The geom to use. Defaults to polygon}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. 
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? \code{NA}, the default, 
includes if any aesthetics are mapped. \code{FALSE} never includes, and \code{TRUE} always includes. 
It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than combining with them. 
This is most useful for helper functions that define both data and aesthetics and shouldn't inherit 
behaviour from the default plot specification, e.g. \code{borders()}.}

\item{...}{Other parameters to pass}
}
\value{
Returns a layer with labels at the circle centers.
}
\description{
Met deze stat kunnen aan een circle-pack labels worden toegevoegd aan het centrum. See \code{pack_circles}
}
\examples{
\dontrun{
 
ggplot(data, aes(y = mpg, label = name)) + stat_packcircles() + stat_packcircles_center
}

}
