% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_packcircles.R
\docType{data}
\name{StatPackCircles}
\alias{StatPackCircles}
\title{StatPackCircles}
\format{An object of class \code{StatPackCircles} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatPackCircles
}
\description{
StatPackCircles
}
\keyword{datasets}
