% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_packcircles.R
\docType{data}
\name{StatPackCirclesLabel}
\alias{StatPackCirclesLabel}
\title{StatPackCirclesLabel}
\format{An object of class \code{StatPackCirclesLabel} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.}
\usage{
StatPackCirclesLabel
}
\description{
StatPackCirclesLabel
}
\keyword{datasets}
