% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi_gegevens.R
\name{toon_knmi_stations}
\alias{toon_knmi_stations}
\title{Toon KNMI-stations op kaart}
\usage{
toon_knmi_stations()
}
\value{
Een intractieve leaflet kaart met KNMI-stations
}
\description{
De functie maakt een interactieve leafletkaart van alle KNMI-station met labels
}
\examples{

toon_knmi_stations()

}
