% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knmi_gegevens.R
\name{knmi_neerslag_dag}
\alias{knmi_neerslag_dag}
\alias{knmi_temperatuur_dag}
\alias{knmi_zonuren_dag}
\alias{knmi_straling_dag}
\title{KNMI-gegevens van neerslag, zonneschijn, straling en temperatuur}
\usage{
knmi_neerslag_dag(knmistation = "344")

knmi_temperatuur_dag(knmistation = "344")

knmi_zonuren_dag(knmistation = "344")

knmi_straling_dag(knmistation = "344")
}
\arguments{
\item{knmistation}{Het nummer van het KNMI-station. Via de website zijn de nummers te vinden of via de
functie \code{toon_knmi_stations}.}
}
\value{
Een dataframe met een kolom datum en neerslag in mm
}
\description{
Deze functie haalt de gegevens op, van een specifiek KNMI-station. De functie maakt gebruik van \code{knmi_dag_ruw}.
Voor meer toelichting kan die functie worden geraadpleegd.
}
\section{Functions}{
\itemize{
\item \code{knmi_neerslag_dag}: Neerslag per dag in mm

\item \code{knmi_temperatuur_dag}: Gemiddelde temperatuur in oC per dag

\item \code{knmi_zonuren_dag}: Zonuren en relatieve zonuren per dag

\item \code{knmi_straling_dag}: Totale straling per dag in J/cm2
}}

\examples{

\dontrun{
neerslag_data <- knmi_neerslag_dag()
neerslag_data <- knmi_neerslag_dag(310)

temperatuur_data <- knmi_temperatuur_dag()

zonuren_data <- knmi_zonuren_dag()

straling_data <- knmi_straling_dag()

}
}
